Compat with WP 4.7
The previous behavior of batcache follows the first example described in the [Make blog post](https://make.wordpress.org/core/2016/09/08/wp_hook-next-generation-actions-and-filters/) regarding `WP_Hook`.
I'm not sure why this manual setting of callbacks in `$wp_filter` is necessary; perhaps it was purely unintentional. If that's the case, using `add_action`/`add_filter` are always preferred.