Normalize GET parameter order
Most requests with query parameters bypass Batcache, but for WordPress REST API requests, query parameters are cached as well, but with a different cache for each unique order of query parameters. If we normalize the query parameter order, we can potentially have a higher number of cache hits.

props: @emrikol