Fix batcache ignoring updated version key
We changed $batcache->times to be 1 instead of 2, but didn't change the condition that checks the integer, so it always returned true which prevented caches from being served, a subsequent change tried to workaround it but broke the version check result.

This brings back the regeneration at the original location, and fixes the condition that broke the above.