Cache redirects when max-age=0
It's not uncommon that when redirects are sent, the Cache-Control also has `max-age`. This is respected by Batcache, which isn't a great default. Even when `$batcache->cache_redirects` is enabled, the `max-age=0` takes precedence. This PR changes that behaviour to cache redirects regardless of the `max-age=X` in the Cache Control header.